add rate limiter

+++ cacheengine/src/main/java/org/example/cacheengine/CacheController.java
@@ -1,8 +1,10 @@
 package org.example.cacheengine;
+import jakarta.servlet.http.HttpServletRequest;
 import org.example.CacheService;
 import org.example.LRUCache;
 import org.springframework.web.bind.annotation.*;
 import org.example.LRUCache;
+//import org.example.RateLimiter;
 import org.example.LFUCache;
 import org.example.TTLCache;
 @RestController
@@ -10,21 +12,25 @@ import org.example.TTLCache;
 public class CacheController {
 
     private final CacheService<String,String> cacheService=new CacheService<>(new LRUCache<>(100));
+    private final RateLimiter rateLimiter = new RateLimiter(3, 60);
 
     @GetMapping("/{key}")
-    public String get(@PathVariable String key){
+    public String get(@PathVariable String key,HttpServletRequest request){
+        if(!rateLimiter.isAllowed(request.getRemoteAddr())) return "429 - Too Many Requests";
         String value= cacheService.get(key);
         return value!=null?value:"null";
     }
 
     @PutMapping("/{key}")
-    public String put(@PathVariable String key,@RequestBody String value){
+    public String put(@PathVariable String key, @RequestBody String value, HttpServletRequest request){
+        if(!rateLimiter.isAllowed(request.getRemoteAddr())) return "429 - Too Many Requests";
         cacheService.put(key,value);
         return "stored";
     }
 
     @DeleteMapping("/{key}")
-    public String delete(@PathVariable String key){
+    public String delete(@PathVariable String key,HttpServletRequest request){
+        if(!rateLimiter.isAllowed(request.getRemoteAddr())) return "429 - Too Many Requests";
         cacheService.delete(key);
         return "deleted";
     }
@@ -35,7 +41,8 @@ public class CacheController {
     }
 
     @PostMapping("/config")
-    public String switchMapping(@RequestParam String policy,@RequestParam int capacity){
+    public String switchMapping(@RequestParam String policy,@RequestParam int capacity,HttpServletRequest request){
+        if(!rateLimiter.isAllowed(request.getRemoteAddr())) return "429 - Too Many Requests";
         switch(policy.toUpperCase()){
             case "LRU":
                 cacheService.switchPolicy(new LRUCache<>(capacity));
